+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Always attempt a location-based RBAC login at startup.
 - Restyle built-in graphs of optimization/training progress.
 - Make the accelerator-selection box a little less janky.
+- Add Nelder–Mead algorithm for single-objective optimization.
 
 ## v0.1.4